requirements.txt
imageio==2.22.4
numpy>=1.23.2
Pillow==9.4.0
